+++ segtypes/common/data.py
@@ -57,7 +57,7 @@ class CommonSegData(CommonSegCodeSubsegment, CommonSegGroup):
                 f.write(preamble + "\n")
 
             f.write(f".section {self.get_linker_section()}")
-            section_flags = self.get_linker_section_flags()
+            section_flags = self.get_section_flags()
             if section_flags:
                 f.write(f', "{section_flags}"')
             f.write("\n\n")